+++ tsdat/transform/converters/bin_average.py
@@ -1,0 +1,5 @@
+from ._adi_base_transformer import _ADIBaseTransformer
+
+
+class BinAverage(_ADIBaseTransformer):
+    transformation_type: str = "TRANS_BIN_AVERAGE"
