+++ tsdat/transform/converters/interpolate.py
@@ -1,0 +1,5 @@
+from ._adi_base_transformer import _ADIBaseTransformer
+
+
+class Interpolate(_ADIBaseTransformer):
+    transformation_type: str = "TRANS_INTERPOLATE"
